Moved files to module, examples now a sub module
