+++ raytracing/examples/confocalDetection.py
