+++ raytracing/examples/ex01.py
@@ -4,7 +4,6 @@ Here we have a single lens of focal length f=50 mm. An object
 (in blue) is positionned at 100 mm in front of the lens. The image
 conjugate (in red) is seen 100 mm after the lens, as expected.
 """
-DIFFICULTY  = 1
 
 from raytracing import *
