+++ raytracing/examples/ex14.py
@@ -1,0 +1,25 @@
+TITLE       = ""
+DESCRIPTION = """
+"""
+
+from raytracing import *
+
+def exempleCode():
+    # Demo #14: Generic objectives
+    obj = Objective(f=10, NA=0.8, focusToFocusLength=60, backAperture=18, workingDistance=2,
+                    magnification=40, fieldNumber=1.4, label="Objective")
+    print("Focal distances: ", obj.focalDistances())
+    print("Position of PP1 and PP2: ", obj.principalPlanePositions(z=0))
+    print("Focal spots positions: ", obj.focusPositions(z=0))
+    print("Distance between entrance and exit planes: ", obj.L)
+
+    path = ImagingPath()
+    path.label = "Demo #14 Path with generic objective"
+    path.append(Space(180))
+    path.append(obj)
+    path.append(Space(10))
+    path.displayWithObject(diameter=20, fanAngle=0.005, comments=DESCRIPTION)
+
+
+if __name__ == "__main__":
+    exempleCode()
