+++ raytracing/examples/ex18.py
@@ -1,0 +1,25 @@
+TITLE       = "Laser beam and vendor lenses"
+DESCRIPTION = """
+"""
+
+from raytracing import *
+
+def exempleCode():
+    # Demo #18: Laser beam and vendor lenses
+    path = LaserPath()
+    path.label = TITLE
+    path.append(Space(d=50))
+    path.append(thorlabs.AC254_050_A())
+    path.append(Space(d=50))
+    path.append(thorlabs.AC254_050_A())
+    path.append(Space(d=150))
+    path.append(eo.PN_33_921())
+    path.append(Space(d=50))
+    path.append(eo.PN_88_593())
+    path.append(Space(d=180))
+    path.append(olympus.LUMPlanFL40X())
+    path.append(Space(d=10))
+    path.display(beams=[GaussianBeam(w=0.001)], comments=DESCRIPTION)
+
+if __name__ == "__main__":
+    exempleCode()
