+++ raytracing/examples/illuminator.py
