+++ raytracing/examples/invariant.py
