+++ raytracing/examples/invariantAndEfficiency.py
