+++ raytracing/examples/kohler.py
