+++ raytracing/examples/kohlerAndInvariant.py
