+++ raytracing/examples/laserScanningMicroscope.py
