+++ raytracing/examples/lsmConfocalpinhole.py
