+++ raytracing/examples/oldexemple.py
