+++ raytracing/examples/randomLambertianRays.py
