+++ raytracing/examples/sourceCollection.py
