+++ raytracing/examples/template.py
@@ -1,0 +1,12 @@
+TITLE       = ""
+DESCRIPTION = """
+"""
+
+from raytracing import *
+
+def exempleCode():
+
+    path.display(comments=DESCRIPTION)
+
+if __name__ == "__main__":
+    exempleCode()
