+++ raytracing/examples/widefield.py
