+++ raytracing/utils.py
@@ -38,7 +38,6 @@ def areAbsolutelyNotEqual(left, right, epsilon=1e-3):
 def areRelativelyNotEqual(left, right, epsilon=1e-3):
     return not areRelativelyAlmostEqual(left, right, epsilon)
 
-
 def deprecated(reason: str):
     def deprecatedFunc(func):
         def wrapper(*args, **kwargs):
